make it all one single dynamic route
